requirements.txt
torch~=2.9.0+cu126
torchvision~=0.24.0+cu126
pillow~=11.1.0